Move inline module to its own file
